Linting,  typechecking and whitlabel fixes (#90)
This PR aims to update notifications UI to be easier to theme / white-label. It was an oversight on my part when doing the dev work not to fully incorporate all white-label items, in the same vein 5.0 has changed with white-labelling (pathing etc).

Would you be able to look at this and see if I have missed anything?

I'm also concerned that I have any breaking changes in here and would like to test this out if we already have a themed version of notifications that we work with or to know if any clients have themed notifications?

Testing is in two halves:

Visually check nothing looks out of the ordinary/different (I may end up doing some before and after screenshots)
Add a theme or use an existing theme if you need one I have you that can be used.

N.B. this also fixes any outstanding ts-errors and adds linting + npm script to check for typescript errors.